Fix docker_host_user_id can't work with docker compose (#83)
Problem:
The script puts the '-e LOCAL_USER_ID=' prefix in the DOCKER_HOST_USER_ID var.

This works fine if you are only using the var in "docker run" commands, but it prevents you from using it in a docker compose file because you need just the user id and not the prefix:

services:
  web:
    environment:
      LOCAL_USER_ID: 0
      PHP_FPM_USER: ${DOCKER_HOST_USER_ID}
      PHP_FPM_LISTEN__OWNER: ${DOCKER_HOST_USER_ID}

Additionally, the var is empty on non-linux machines, which is actually fine because it's not needed. But this adds complexity because it is not trivial to omit the env vars from docker compose files. It is much easier to have the var declared even when it is not needed.

Fix:
Modify the script to only supply the user id. Devs will need to add the prefix where necessary.
Add a user id for non-linux machines for the sake of simplicity.