Fix relative paths
The paths were not update to reflect that the scripts are no longer being executed via a symlink.

As a result, the relative path was off. e.g. the relative path from a './bin/cs-style-fixer' symlink to the root is '../', whereas from a non-symlink would be './submodules/bin/cs-style-fixer' with a relative path to the root of '../../'