Fix pre-commit fails if no files are staged
Problem:
If you try to commit with no staged files, e.g. using "git commit --amend" to edit the commit message, then you are returned to the command line with a non-zero exit code and not allowed to ammend the commit as expected.

Cause:
The pre-commit script in submodules is now using -o pipefail, which is causing the script to exit when the grep in the staged_files sub-shell returns a non-zero exit code when it does not find any matches.

Fix:
Temporarily disable the -e option when using this subshell.