Fix ~/.ssh permission errors with hub (#103)
Problem:
When running hub on Mac OS X, the hub command errors out with:

Cause:
The hub command is using the user_ssh Docker volume to mount the SSH dir. Mac OS X uses NFS for the Docker volumes, which presents the true owner, group and permissions to the NFS client. Since the user in the hub container is not the same user on the Mac OS X Docker host, this problem occurrs.

Fix:
Do not use Docker volumes for the ~/.ssh mount. Instead, do a direct mount with -v. On Mac OS X, this will result in Docker's normal behavior where it shows the current Linux user as the owner of the mount, regardless of who the true owner on the host is. On Linux, we as the user id of the Docker container user to match the host, so even though the true owner is visible, the user id will match, so all is good there too.