Add executable script for calling circleci.job.queue
Problem:
The circleci.job.queue function is part of the circleci.sh library. Which is great for sharing it, but introduces some difficult when wanting to invoke it:

1. circleci.sh is a Bash script. If you source it from a non-bash shell, certain things will not work, such as the BASH_SOURCE env var.
2. If you source circleci.sh to access the functions, you get `set -euo pipefail` affecting your current shell session, which is not necessarily desired.

Fix:
In addition to allowing the file to be sourced when that is the desired use-case, also include an executable Bash script that will allow the function to be called while isolating the caller's shell from the effects of the script.