Pass Netlify env vars to node container (#109)
Problem:
Projects that use Netlify, such as [ng-checkout](https://github.com/Medology/ng-checkout) are failing to deploy Netlify previews when using the node container from Submodules.

Cause:
To deploy a Netlify preview, two env vars are needed: NETLIFY_AUTH_TOKEN & NETLIFY_SITE_ID. These are not being passed to the container.

Fix:
Pass the env vars to the container.

Note:
Sincethe env vars are being passed by name, it will not cause a problem if the env vars are not set because we are not referencing them directly in the shell script. Instead, Docker will inspect the current environment and if there is a matching env var defined, the value will be passed to the container. If there is not a matching env var defined, Docker will just ignore it.